feature: Separate benchmarks
